renamed and cleaned up example for custom variables dialog
